add: exit handling to executor
