--- _strcmp.c
@@ -1,28 +1,0 @@
-#include "main.h"
-
-/**
- * _strcmp - compares two strings
- * @s1: first string
- * @s2: second string
- *
- * Return: 1 if stings are the same, 0 otherwise
- */
-
-int _strcmp(char *s1, char *s2)
-{
-	int i = 0;
-
-	while (1)
-	{
-		if (s1[i] != s2[i])
-			return (0);
-
-		if (s1[i] == 0 || s2[i] == 0)
-			return (1);
-
-		i++;
-	}
-
-	return (1);
-}
-
